ur_robot_driver builds on windows

+++ ur_robot_driver/include/ur_robot_driver/comm/stream.h
@@ -19,8 +19,7 @@
  */
 
 #pragma once
-#include <netdb.h>
-#include <sys/socket.h>
+
 #include <sys/types.h>
 #include <atomic>
 #include <mutex>
@@ -107,7 +106,7 @@ public:
 protected:
   virtual bool open(int socket_fd, struct sockaddr* address, size_t address_len)
   {
-    return ::connect(socket_fd, address, address_len) == 0;
+    return ::connect(socket_fd, address, static_cast<int>(address_len)) == 0;
   }
 
 private:
