+++ ur_robot_driver/include/ur_robot_driver/rtde/package_header.h
@@ -30,7 +30,7 @@
 #define UR_RTDE_DRIVER_RTDE__HEADER_H_INCLUDED
 
 #include <cstddef>
-#include <endian.h>
+#include <ur_robot_driver/portable_endian.h>
 #include "ur_robot_driver/types.h"
 #include "ur_robot_driver/comm/package_serializer.h"
 
@@ -84,10 +84,10 @@ public:
    *
    * \returns
    */
-  static size_t serializeHeader(uint8_t* buffer, PackageType package_type, uint16_t payload_length)
+  static size_t serializeHeader(uint8_t* buffer, PackageType package_type, size_t payload_length)
   {
-    uint16_t header_size = sizeof(_package_size_type) + sizeof(PackageType);
-    uint16_t size = header_size + payload_length;
+    size_t header_size = sizeof(_package_size_type) + sizeof(PackageType);
+    size_t size = header_size + payload_length;
     comm::PackageSerializer::serialize(buffer, size);
     comm::PackageSerializer::serialize(buffer + sizeof(size), package_type);
     return header_size;
